优化云端录像转化

+++ sql/初始化.sql
@@ -282,7 +282,7 @@ create table wvp_cloud_record (
                                   collect bool default false,
                                   reserve bool default false,
                                   file_size integer,
-                                  time_len integer,
+                                  time_len float,
                                   constraint uk_stream_push_app_stream_path unique (app, stream, file_path)
 );
